Improve the organization of tests
