+++ src/metrics/graph_metrics.rs
@@ -293,6 +293,32 @@ fn bfs_distances<A, W, Ty: GraphConstructor<A, W> + EdgeType>(
 
 #[cfg(test)]
 mod tests {
+
+    #[test]
+    fn test_assortativity_is_symmetric_newman_coefficient() {
+        use crate::core::types::Graph;
+        // Degree assortativity is the Pearson correlation over the symmetric joint
+        // degree distribution of edge endpoints: each undirected edge contributes
+        // both orderings. A star is perfectly disassortative (a degree-k hub joined
+        // only to degree-1 leaves), so its coefficient is exactly -1.0. The earlier
+        // implementation correlated a single edge ordering, giving the two endpoints
+        // different means; on a star that collapsed the variance to zero and
+        // returned 0.0 instead of -1.0.
+        use crate::metrics::assortativity;
+
+        let mut g: Graph<i32, f64> = Graph::new();
+        let center = g.add_node(0);
+        for i in 1..=4 {
+            let leaf = g.add_node(i);
+            g.add_edge(center, leaf, 1.0);
+        }
+
+        let r = assortativity(&g);
+        assert!(
+            (r - (-1.0)).abs() < 1e-9,
+            "star degree assortativity should be -1.0, got {r}"
+        );
+    }
     use super::*;
     use crate::core::types::Graph;
